+++ src/app/Console/Commands/Update.php
@@ -37,17 +37,27 @@ class Update extends Command
         $vendorscss = base_path() . '/vendor/pveltrop/dcms/src/app/Resources/sass';
         $vendorcssassets = base_path() . '/vendor/pveltrop/dcms/src/app/Public/css/assets';
 
-        RemoveDir($vendorjs);
-        RemoveDir($vendorjsassets);
-        RemoveDir($vendorscss);
-        RemoveDir($vendorcssassets);
+        if ($console->confirm('Do you want to update the JavaScript files?')){
+            // JS
+            RemoveDir($vendorjs);
+            CopyDir($rootjs,$vendorjs);
+            RemoveDir($vendorjsassets);
+            CopyDir($rootjsassets,$vendorjsassets);
+        }
 
-        CopyDir($rootjs,$vendorjs);
-        CopyDir($rootjsassets,$vendorjsassets);
-        CopyDir($rootscss,$vendorscss);
-        CopyDir($rootcssassets,$vendorcssassets);
-        copy(base_path() . '/config/dcms.php', base_path() . '/vendor/pveltrop/dcms/src/Config/dcms.php');
-        copy(base_path() . '/dcms.json', base_path() . '/vendor/pveltrop/dcms/src/dcms.json');
+        if ($console->confirm('Do you want to update the CSS files?')){
+            // CSS        
+            RemoveDir($vendorscss);
+            CopyDir($rootscss,$vendorscss);
+            RemoveDir($vendorcssassets);
+            CopyDir($rootcssassets,$vendorcssassets);
+        }
+        
+        if ($console->confirm('Do you want to update the configs?')){
+            // Configs
+            copy(base_path() . '/config/dcms.php', base_path() . '/vendor/pveltrop/dcms/src/Config/dcms.php');
+            copy(base_path() . '/dcms.json', base_path() . '/vendor/pveltrop/dcms/src/dcms.json');
+        }
 
         $console->comment('');
         $console->comment('Updated DCMS vendor package. Create a merge request to propose changes.');
